+++ interfaces/global-state-types.ts
@@ -12,4 +12,5 @@ export interface GlobalStateInterface {
   hasReports?: boolean;
   openSearch?: boolean;
   dismissFeaturedInsights?: boolean;
+  workspaces: Workspace[];
 }